Make each admin protocol individually loadable
And add groups to load multiple based on category.
