+++ acapy_plugin_toolbox/connections.py
@@ -5,6 +5,8 @@
 
 from marshmallow import fields, validate
 
+from aries_cloudagent.config.injection_context import InjectionContext
+from aries_cloudagent.core.protocol_registry import ProtocolRegistry
 from aries_cloudagent.messaging.base_handler import BaseHandler, BaseResponder, RequestContext
 from aries_cloudagent.protocols.connections.manager import ConnectionManager
 from aries_cloudagent.connections.models.connection_record import (
@@ -82,6 +84,14 @@ MESSAGE_TYPES = {
         'acapy_plugin_toolbox.connections'
         '.UpdateConnection',
 }
+
+
+async def setup(context: InjectionContext):
+    """Setup the connections plugin."""
+    protocol_registry = await context.inject(ProtocolRegistry)
+    protocol_registry.register_message_types(
+        MESSAGE_TYPES
+    )
 
 
 ConnectionGetList, ConnectionGetListSchema = generate_model_schema(
