+++ acapy_plugin_toolbox/credential_definitions.py
@@ -7,6 +7,8 @@ from asyncio import shield
 
 from marshmallow import fields
 
+from aries_cloudagent.config.injection_context import InjectionContext
+from aries_cloudagent.core.protocol_registry import ProtocolRegistry
 from aries_cloudagent.messaging.base_handler import BaseHandler, BaseResponder, RequestContext
 from aries_cloudagent.messaging.models.base_record import BaseRecord, BaseRecordSchema
 from aries_cloudagent.protocols.problem_report.message import ProblemReport
@@ -41,6 +43,15 @@ MESSAGE_TYPES = {
     CRED_DEF_LIST:
         'acapy_plugin_toolbox.credential_definitions.CredDefList',
 }
+
+
+
+async def setup(context: InjectionContext):
+    """Setup the cred def plugin."""
+    protocol_registry = await context.inject(ProtocolRegistry)
+    protocol_registry.register_message_types(
+        MESSAGE_TYPES
+    )
 
 
 class CredDefRecord(BaseRecord):
