+++ acapy_plugin_toolbox/dids.py
@@ -5,6 +5,9 @@
 
 from typing import Dict
 from marshmallow import fields
+
+from aries_cloudagent.config.injection_context import InjectionContext
+from aries_cloudagent.core.protocol_registry import ProtocolRegistry
 from aries_cloudagent.messaging.base_handler import BaseHandler, BaseResponder, RequestContext
 from aries_cloudagent.messaging.models.base_record import BaseRecord, BaseRecordSchema
 from aries_cloudagent.wallet.base import BaseWallet, DIDInfo
@@ -56,6 +59,14 @@ MESSAGE_TYPES = {
     'acapy_plugin_toolbox.dids'
         '.GetDidEndpoint'
 }
+
+
+async def setup(context: InjectionContext):
+    """Setup the dids plugin."""
+    protocol_registry = await context.inject(ProtocolRegistry)
+    protocol_registry.register_message_types(
+        MESSAGE_TYPES
+    )
 
 
 class DidRecord(BaseRecord):
