+++ acapy_plugin_toolbox/schemas.py
@@ -7,6 +7,8 @@ from asyncio import shield
 
 from marshmallow import fields
 
+from aries_cloudagent.config.injection_context import InjectionContext
+from aries_cloudagent.core.protocol_registry import ProtocolRegistry
 from aries_cloudagent.messaging.base_handler import BaseHandler, BaseResponder, RequestContext
 from aries_cloudagent.messaging.models.base_record import BaseRecord, BaseRecordSchema
 from aries_cloudagent.ledger.base import BaseLedger
@@ -44,6 +46,14 @@ MESSAGE_TYPES = {
         'acapy_plugin_toolbox.schemas'
         '.SchemaList',
 }
+
+
+async def setup(context: InjectionContext):
+    """Setup the schemas plugin."""
+    protocol_registry = await context.inject(ProtocolRegistry)
+    protocol_registry.register_message_types(
+        MESSAGE_TYPES
+    )
 
 
 class SchemaRecord(BaseRecord):
